[REN][9.0] project_timesheet_time_control: Module rename and fix suggested by @pedrobaeza
