+++ project_timesheet_time_control/__openerp__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+# © 2016 Antonio Espinosa
+# © 2016 Pedro M. Baeza
+# © 2016 Sergio Teruel
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+{
+    'name': 'Project timesheet time control',
+    'version': '9.0.1.0.0',
+    'category': 'Project',
+    'author': 'Serv. Tecnol. Avanzados - Pedro M. Baeza, '
+              'Antiun Ingeniería S.L.',
+    'website': 'http://www.serviciosbaeza.com',
+    'depends': ['project_timesheet'],
+    'data': [
+        'security/project_security.xml',
+        'view/account_analytic_line_view.xml',
+    ],
+    'installable': True,
+}
