+++ project_timesheet_time_control/i18n/es.po
@@ -1,0 +1,82 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* project_work_time_control
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 9.0c\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-03-21 17:58+0000\n"
+"PO-Revision-Date: 2016-03-21 19:01+0100\n"
+"Last-Translator: Sergio Teruel <sergio@incaser.es>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: es_ES\n"
+"X-Generator: Poedit 1.5.4\n"
+
+#. module: project_work_time_control
+#: model:ir.model,name:project_work_time_control.model_account_analytic_line
+msgid "Analytic Line"
+msgstr "Línea analítica"
+
+#. module: project_work_time_control
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_timesheet_account_analytic_line_tree
+msgid "Close task"
+msgstr "Cerrar tarea"
+
+#. module: project_work_time_control
+#: model:ir.model.fields,field_description:project_work_time_control.field_account_analytic_line_date_time
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_analytic_line_search
+msgid "Date"
+msgstr "Fecha"
+
+#. module: project_work_time_control
+#: model:ir.model.fields,field_description:project_work_time_control.field_account_analytic_line_folded
+msgid "Folded in Tasks Pipeline"
+msgstr "Replegado en la vista Kanban"
+
+#. module: project_work_time_control
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_analytic_line_search
+msgid "Group By..."
+msgstr "Agrupar por..."
+
+#. module: project_work_time_control
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_timesheet_account_analytic_line_tree
+msgid "Open task"
+msgstr "Abrir tarea"
+
+#. module: project_work_time_control
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_analytic_line_search
+#, fuzzy
+msgid "Search task work"
+msgstr "Actualizar Modo Trabajo"
+
+#. module: project_work_time_control
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_timesheet_account_analytic_line_tree
+msgid "Stop"
+msgstr "Stop"
+
+#. module: project_work_time_control
+#: model:ir.model,name:project_work_time_control.model_project_task
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_analytic_line_search
+msgid "Task"
+msgstr "Tarea"
+
+#. module: project_work_time_control
+#: model:ir.model,name:project_work_time_control.model_project_task_type
+msgid "Task Stage"
+msgstr "Estado actual"
+
+#. module: project_work_time_control
+#: model:ir.model.fields,help:project_work_time_control.field_account_analytic_line_folded
+msgid ""
+"This stage is folded in the kanban view when there are no records in that "
+"stage to display."
+msgstr ""
+
+#. module: project_work_time_control
+#: model:ir.ui.view,arch_db:project_work_time_control.work_time_control_analytic_line_search
+msgid "User"
+msgstr "Usuario"
