+++ project_timesheet_time_control/view/account_analytic_line_view.xml
@@ -2,8 +2,8 @@
 <openerp>
     <data>
 
-       <record id="work_time_control_timesheet_account_analytic_line_tree" model="ir.ui.view">
-            <field name="name">work.time.control.account.analytic.line.tree</field>
+       <record id="account_analytic_line_tree" model="ir.ui.view">
+            <field name="name">project.timesheet.time.control.account.analytic.line.tree</field>
             <field name="model">account.analytic.line</field>
             <field name="inherit_id"
                    ref="project_timesheet.view_account_analytic_line_tree_inherit_account_id"/>
@@ -20,24 +20,24 @@
                     />
                 </field>
                 <field name="task_id" position="after">
-                    <field name="include_in_task_work_view" invisible="True"/>
+                    <field name="folded" invisible="1"/>
                     <button name="button_close_task"
                             string="Close task"
                             type="object"
                             icon="gtk-apply"
-                            attrs="{'invisible': [('include_in_task_work_view', '=', False)]}"
+                            attrs="{'invisible': [('folded', '=', True)]}"
                             />
                     <button name="button_open_task"
                             string="Open task"
                             type="object"
                             icon="gtk-redo"
-                            attrs="{'invisible': [('include_in_task_work_view', '=', True)]}"
+                            attrs="{'invisible': [('folded', '=', False)]}"
                             />
                 </field>
             </field>
        </record>
 
-        <record id="work_time_control_analytic_line_search" model="ir.ui.view">
+        <record id="account_analytic_line_search" model="ir.ui.view">
             <field name="name">account.analytic.line.search</field>
             <field name="model">account.analytic.line</field>
             <field name="arch" type="xml">
