--- project_work_time_control/README.rst
@@ -1,20 +1,0 @@
-Fast encoding of project task works
-===================================
-
-Adds a button at task work level to compute minutes lasted from start date to
-the current moment.
-
-It also adds a direct menu entry for task works fast encoding.
-
-Usage
-=====
-To use this module, go to Project > Project > Tasks works, and you can add
-easily works of your tasks there.
-
-Credits
-=======
-
-Contributors
-------------
-* Pedro M. Baeza <pedro.baeza@serviciosbaeza.com>
-* Antonio Espinosa <antonioea@antiun.com>
