--- project_work_time_control/i18n/es.po
@@ -1,59 +1,0 @@
-# Translation of OpenERP Server.
-# This file contains the translation of the following modules:
-#	* project_work_time_control
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: OpenERP Server 7.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-01-28 09:06+0000\n"
-"PO-Revision-Date: 2015-01-28 09:06+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: project_work_time_control
-#: view:project.task.work:project_work_time_control.project_task_work_search
-msgid "Date"
-msgstr "Fecha"
-
-#. module: project_work_time_control
-#: view:project.task.work:project_work_time_control.project_task_work_search
-msgid "Group By..."
-msgstr "Agrupar por..."
-
-#. module: project_work_time_control
-#: help:project.task.type,include_in_task_work_view:0
-msgid "If you mark this check, tasks that are in this stage will be selectable in the 'Tasks works' fast-encoding view."
-msgstr "Si marca esta casilla, las tareas que estén en esta etapa serán seleccionables en la vista de codificación rápida 'Trabajos de las tareas'."
-
-#. module: project_work_time_control
-#: field:project.task.type,include_in_task_work_view:0
-msgid "Include in the task work fast-encoding view"
-msgstr "Incluir en la vista de codificación rápida de trabajos de tareas"
-
-#. module: project_work_time_control
-#: field:project.task.work,project:0
-msgid "Project"
-msgstr "Proyecto"
-
-#. module: project_work_time_control
-#: view:project.task:0
-#: view:project.task.work:0
-msgid "Stop"
-msgstr "Parar"
-
-#. module: project_work_time_control
-#: view:project.task.work:0
-msgid "Task work"
-msgstr "Trabajo de tarea"
-
-#. module: project_work_time_control
-#: model:ir.actions.act_window,name:project_work_time_control.project_task_work_action
-#: model:ir.ui.menu,name:project_work_time_control.project_task_work_menu
-msgid "Tasks works"
-msgstr "Trabajos de las tareas"
-
