--- project_work_time_control/view/project_task_type_view.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record model="ir.ui.view" id="task_type_edit_time">
-            <field name="name">project.task.type.form.time</field>
-            <field name="model">project.task.type</field>
-            <field name="inherit_id" ref="project.task_type_edit" />
-            <field name="arch" type="xml">
-                <field name="fold" position="after">
-                    <field name="include_in_task_work_view"/>
-                </field>
-            </field>
-        </record>
-
-    </data>
-</openerp>
-
